+++ _unittests/ut_numbers/test_weighted_double.py
@@ -1,13 +1,14 @@
 """
-@brief      test log(time=28s)
+@brief      test log(time=2s)
 """
 
 
 import sys
 import os
 import unittest
+from inspect import signature, isbuiltin, isclass, ismethod, isfunction, _signature_fromstr, Signature
 from pyquickhelper.pycode import ExtTestCase
-from inspect import signature
+from pyquickhelper.helpgen import rst2html
 
 
 try:
@@ -29,7 +30,6 @@ from src.cpyquickhelper.numbers import WeightedDouble, WeightedFloat
 class TestWeightedDouble(ExtTestCase):
 
     def test_creation(self):
-
         n1 = WeightedDouble(1, 1)
         n2 = WeightedDouble(1, 1)
         n3 = n1 + n2
@@ -53,7 +53,28 @@ class TestWeightedDouble(ExtTestCase):
         self.assertEqual(n3, WeightedFloat(2, 2))
 
     def test_signature(self):
-        self.assertRaise(lambda: signature(WeightedDouble.__init__), ValueError)
+        self.assertRaise(lambda: signature(
+            WeightedDouble.__init__), ValueError)
+        self.assertFalse(isbuiltin(WeightedDouble))
+        self.assertFalse(isbuiltin(WeightedDouble.__init__))
+        self.assertTrue(isclass(WeightedDouble))
+        self.assertFalse(ismethod(WeightedDouble.__init__))
+        self.assertFalse(isfunction(WeightedDouble.__init__))
+        sig = "__init__(value, weight=1)"
+        res = _signature_fromstr(Signature, WeightedDouble.__init__, sig)
+        self.assertNotEmpty(res)
+        newstring = [
+            ".. autosignature:: src.cpyquickhelper.numbers.weighted_number.WeightedDouble.__init__"]
+        newstring = "\n".join(newstring)
+        res = rst2html(newstring, writer="rst", layout="sphinx")
+        self.assertIn(
+            "src.cpyquickhelper.numbers.weighted_number.WeightedDouble.__init__", res)
+        self.assertIn("(*self*, *value*, *weight* = *1.0*)", res)
+
+    def test_attribute(self):
+        n = WeightedDouble(4, 5)
+        self.assertEqual(n.value, 4)
+        self.assertEqual(n.weight, 5)
 
 
 if __name__ == "__main__":