+++ src/cpyquickhelper/numbers/__init__.py
@@ -3,4 +3,5 @@
 @brief Shortcut to *numbers*.
 """
 
+from .weighted_dataframe import WeightedSeries
 from .weighted_number import WeightedDouble, WeightedFloat