+++ src/cpyquickhelper/numbers/weighted_number_python.cpp
@@ -11,10 +11,16 @@ PYBIND11_MODULE(weighted_number, m) {
 	py::class_<WeightedDouble>(m, "WeightedDouble",
 			R"pbdoc(
 			Implements a weighted double used to speed up computation with aggregation.
+			It contains two attributes:
+			* value: unweighted value
+			* weight: weight associated to the value, it should be positive,
+              but that's not enforced
 			)pbdoc")
 		.def(py::init<double, double>(), py::arg("value"), py::arg("weight") = 1.)
-		.def("__str__", &WeightedDouble::__str__)
-		.def("__repr__", &WeightedDouble::__repr__)
+		.def("__str__", &WeightedDouble::__str__, "usual")
+		.def("__repr__", &WeightedDouble::__repr__, "usual")
+		.def_readwrite("value", &WeightedDouble::value, "unweighted numeric value (counter)")
+		.def_readwrite("weight", &WeightedDouble::weight, "weight")
 		.def(py::self + py::self)
 		.def(py::self - py::self)
 		.def(py::self * py::self)
@@ -43,10 +49,16 @@ PYBIND11_MODULE(weighted_number, m) {
 	py::class_<WeightedFloat>(m, "WeightedFloat",
 		R"pbdoc(
 			Implements a weighted float used to speed up computation with aggregation.
+			It contains two attributes:
+			* value: unweighted value
+			* weight: weight associated to the value, it should be positive,
+              but that's not enforced
 			)pbdoc")
 		.def(py::init<float, float>(), py::arg("value"), py::arg("weight") = 1.)
-		.def("__str__", &WeightedFloat::__str__)
-		.def("__repr__", &WeightedFloat::__repr__)
+		.def("__str__", &WeightedFloat::__str__, "usual")
+		.def("__repr__", &WeightedFloat::__repr__, "usual")
+		.def_readwrite("value", &WeightedFloat::value, "unweighted numeric value (counter)")
+		.def_readwrite("weight", &WeightedFloat::weight, "weight")
 		.def(py::self + py::self)
 		.def(py::self - py::self)
 		.def(py::self * py::self)