+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 current - 2018-08-03 - 0.00Mb
 =============================
 
+* `4`: implements a series with WeightedDouble (2018-08-03)
 * `3`: add an example based on pybind11 (2018-08-02)
 * `2`: add history (2018-04-01)